+++ yash-env/src/pwd.rs
@@ -143,7 +143,7 @@ mod tests {
     }
 
     fn env_with_symlink_to_dir() -> Env<VirtualSystem> {
-        let mut system = VirtualSystem::new();
+        let system = VirtualSystem::new();
         let mut state = system.state.borrow_mut();
         state
             .file_system
@@ -246,7 +246,7 @@ mod tests {
 
     #[test]
     fn prepare_pwd_with_non_absolute_path() {
-        let mut system = VirtualSystem::new();
+        let system = VirtualSystem::new();
         let mut state = system.state.borrow_mut();
         state
             .file_system