+++ yash-env/CHANGELOG.md
@@ -46,6 +46,9 @@ A _private dependency_ is used internally and not visible to downstream users.
   synchronous `Result<()>`. This change allows virtual systems to simulate the
   blocking behavior of `tcsetpgrp` when called from a background process group,
   though the current virtual system implementation does not yet do so.
+- `system::virtual::VirtualSystem::current_process_mut` now takes `&self`
+  instead of `&mut self` because it returns a `std::cell::RefMut` out of a
+  `RefCell`.
 - `impl<S> std::fmt::Debug for builtin::Builtin<S>` now prints the `execute`
   field as well.